+++ plugins/app-json/.gitignore
@@ -2,3 +2,4 @@
 /triggers
 /pre-*
 /post-*
+/install
