+++ plugins/app-json/Makefile
@@ -1,4 +1,4 @@
-TRIGGERS = triggers/post-deploy triggers/pre-deploy
+TRIGGERS = triggers/install triggers/post-delete triggers/post-deploy triggers/pre-deploy
 BUILD = triggers
 PLUGIN_NAME = app-json
