+++ plugins/cron/Makefile
@@ -1,0 +1,6 @@
+SUBCOMMANDS = subcommands/list subcommands/report
+TRIGGERS = triggers/post-delete triggers/post-deploy triggers/report
+BUILD = commands subcommands triggers
+PLUGIN_NAME = cron
+
+include ../../common.mk
