+++ plugins/cron/go.mod
@@ -1,0 +1,15 @@
+module github.com/dokku/dokku/plugins/cron
+
+go 1.15
+
+require (
+	github.com/dokku/dokku/plugins/app-json v0.0.0-00010101000000-000000000000
+	github.com/dokku/dokku/plugins/common v0.0.0-00010101000000-000000000000
+	github.com/robfig/cron/v3 v3.0.1
+	github.com/ryanuber/columnize v1.1.2-0.20190319233515-9e6335e58db3
+	github.com/spf13/pflag v1.0.5 // indirect
+)
+
+replace github.com/dokku/dokku/plugins/app-json => ../app-json
+
+replace github.com/dokku/dokku/plugins/common => ../common
