+++ plugins/cron/src/triggers/triggers.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"strings"
 
-	appjson "github.com/dokku/dokku/plugins/app-json"
 	"github.com/dokku/dokku/plugins/common"
+	cron "github.com/dokku/dokku/plugins/cron"
 )
 
 // main entrypoint to all triggers
@@ -18,19 +18,13 @@ func main() {
 
 	var err error
 	switch trigger {
-	case "install":
-		err = appjson.TriggerInstall()
 	case "post-delete":
-		appName := flag.Arg(0)
-		err = appjson.TriggerPostDelete(appName)
+		err = cron.TriggerPostDelete()
 	case "post-deploy":
+		err = cron.TriggerPostDeploy()
+	case "report":
 		appName := flag.Arg(0)
-		imageTag := flag.Arg(3)
-		err = appjson.TriggerPostDeploy(appName, imageTag)
-	case "pre-deploy":
-		appName := flag.Arg(0)
-		imageTag := flag.Arg(1)
-		err = appjson.TriggerPreDeploy(appName, imageTag)
+		err = cron.ReportSingleApp(appName, "")
 	default:
 		common.LogFail(fmt.Sprintf("Invalid plugin trigger call: %s", trigger))
 	}
