+++ plugins/scheduler-docker-local/scheduler-run
@@ -37,6 +37,10 @@ trigger-scheduler-docker-local-scheduler-run() {
   DOCKER_ARGS+=$(: | plugn trigger docker-args-process-run "$APP" "$IMAGE_SOURCE_TYPE" "$IMAGE_TAG")
   DOCKER_ARGS+=" -e DYNO=run.$DYNO_NUMBER --name $APP.run.$DYNO_NUMBER"
 
+  if [[ -n "$DOKKU_CRON_ID" ]]; then
+    DOCKER_ARGS+=" --label=com.dokku.cron-id=$DOKKU_CRON_ID"
+  fi
+
   declare -a ARG_ARRAY
   eval "ARG_ARRAY=($DOCKER_ARGS)"
