+++ tests/apps/python/task.py
@@ -1,0 +1,5 @@
+def main(args):
+    print(args)
+
+if __name__ == '__main__':
+    main(sys.argv)
